refactor: fold ghost-mcp into ghost-ui
ghost-mcp had no identity independent of ghost-ui — its only job was
re-exposing ghost-ui's registry.json over the MCP transport. Move the
six source files into packages/ghost-ui/src/mcp/ and expose the
"ghost-mcp" bin from ghost-ui directly.
