+++ .changeset/config.json
@@ -7,5 +7,5 @@
   "access": "public",
   "baseBranch": "main",
   "updateInternalDependencies": "patch",
-  "ignore": ["ghost-ui", "ghost-mcp", "ghost-docs"]
+  "ignore": ["ghost-ui", "ghost-docs"]
 }
