--- packages/ghost-mcp/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "ghost-mcp",
-  "version": "0.1.0",
-  "private": true,
-  "description": "MCP server for the Ghost UI component registry",
-  "license": "Apache-2.0",
-  "type": "module",
-  "bin": {
-    "ghost-mcp": "./dist/bin.js"
-  },
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "tsc --build && node scripts/copy-assets.mjs"
-  },
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.29.0",
-    "zod": "^3.25.0"
-  }
-}
