--- packages/ghost-mcp/src/index.ts
@@ -1,1 +1,0 @@
-export { createServer } from "./server.js";
