+++ packages/ghost-ui/package.json
@@ -6,6 +6,9 @@
   "main": "./dist-lib/index.js",
   "module": "./dist-lib/index.js",
   "types": "./dist-lib/index.d.ts",
+  "bin": {
+    "ghost-mcp": "./dist-mcp/bin.js"
+  },
   "exports": {
     ".": {
       "types": "./dist-lib/index.d.ts",
@@ -18,6 +21,7 @@
   },
   "files": [
     "dist-lib",
+    "dist-mcp",
     "src/styles",
     "registry.json"
   ],
@@ -25,8 +29,9 @@
     "**/*.css"
   ],
   "scripts": {
-    "build": "vite build --config vite.lib.config.ts && tsc --build tsconfig.lib.json && node scripts/resolve-dts-paths.mjs",
+    "build": "vite build --config vite.lib.config.ts && tsc --build tsconfig.lib.json && node scripts/resolve-dts-paths.mjs && tsc --build tsconfig.mcp.json",
     "build:lib": "vite build --config vite.lib.config.ts && tsc --build tsconfig.lib.json && node scripts/resolve-dts-paths.mjs",
+    "build:mcp": "tsc --build tsconfig.mcp.json",
     "build:base": "node scripts/build-base-vars.mjs",
     "build:presets": "node scripts/build-presets.mjs",
     "generate:skills": "node scripts/generate-skills.mjs",
@@ -36,6 +41,7 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^5.2.2",
+    "@modelcontextprotocol/sdk": "^1.29.0",
     "@radix-ui/react-accordion": "^1.2.12",
     "@radix-ui/react-alert-dialog": "^1.1.15",
     "@radix-ui/react-aspect-ratio": "^1.1.8",
