+++ packages/ghost-ui/src/mcp/bin.ts
