+++ packages/ghost-ui/src/mcp/server.ts
