+++ packages/ghost-ui/src/mcp/tools.ts
