+++ tsconfig.json
@@ -14,6 +14,6 @@
   "files": [],
   "references": [
     { "path": "packages/ghost-drift" },
-    { "path": "packages/ghost-mcp" }
+    { "path": "packages/ghost-ui/tsconfig.mcp.json" }
   ]
 }
